+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 - multi-source-job-queue: funnels arbitrary other job queues into a single source
 - "self" field in various log records for correlation
+- config: Initial sleep duration prior to beginning job execution
 
 ### Changed
 - amqp-log-writer, http-log-writer: check context done to prevent goroutine leakage
@@ -15,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
     - attach context to all HTTP requests
     - more debug logging
 - step-write-worker-info: report the job type (amqp/http/file) in instance line
+- processor: check for cancellation in between various steps
 
 ### Deprecated